requirements.txt
opcua
psycopg2-binary
openpyxl
python-dateutil

